+++ backend/api/Cargo.toml
@@ -91,6 +91,7 @@ presentation = {path = "../common/presentation"}
 olog = {path = "../common/olog"}
 
 [dev-dependencies]
+testing = {path = "../common/testing"}
 lazy_static = "1.4.0"
 rstest = "0.15.0"
 mockall = "0.11.2"
