+++ backend/common/domain/src/payment/aggregate.rs
@@ -200,11 +200,15 @@ mod tests {
 
 	use assert_matches::assert_matches;
 	use rstest::{fixture, rstest};
-	use testing::fixtures::payment::constants::{CONTRACT_ADDRESSES, TRANSACTION_HASHES};
 	use uuid::Uuid;
 
 	use super::*;
 	use crate::{BlockchainNetwork, Currency, EthereumAddress, PaymentReceiptId, UserId};
+
+	pub const CONTRACT_ADDRESSES: [&str; 1] = ["0xd8da6bf26964af9d7eed9e03e53415d37aa96045"];
+
+	pub const TRANSACTION_HASHES: [&str; 1] =
+		["0x797fb77202901c52094d2544f3631a3535b8ca40009f6a6ac6940b67e6873a4"];
 
 	#[fixture]
 	fn payment_receipt_id() -> PaymentReceiptId {
