+++ backend/common/domain/src/payment/events.rs
@@ -72,11 +72,15 @@ mod tests {
 	use chrono::{NaiveDate, NaiveTime};
 	use rstest::*;
 	use serde_json::{json, Value};
-	use testing::fixtures::payment::constants::{CONTRACT_ADDRESSES, TRANSACTION_HASHES};
 	use uuid::Uuid;
 
 	use super::*;
 	use crate::{BlockchainNetwork, Currency, EthereumName, TransactionHash};
+
+	pub const CONTRACT_ADDRESSES: [&str; 1] = ["0xd8da6bf26964af9d7eed9e03e53415d37aa96045"];
+
+	pub const TRANSACTION_HASHES: [&str; 1] =
+		["0x797fb77202901c52094d2544f3631a3535b8ca40009f6a6ac6940b67e6873a4"];
 
 	#[fixture]
 	fn payment_id() -> PaymentId {
