+++ backend/common/testing/src/context/mod.rs
@@ -1,0 +1,2 @@
+pub mod amqp;
+pub mod database;
