--- backend/common/testing/src/fixtures/mod.rs
@@ -1,1 +1,0 @@
-pub mod payment;
