--- backend/common/testing/src/fixtures/payment/constants.rs
@@ -1,4 +1,0 @@
-pub const CONTRACT_ADDRESSES: [&str; 1] = ["0xd8da6bf26964af9d7eed9e03e53415d37aa96045"];
-
-pub const TRANSACTION_HASHES: [&str; 1] =
-	["0x797fb77202901c52094d2544f3631a3535b8ca40009f6a6ac6940b67e6873a4"];
