--- backend/common/testing/src/fixtures/payment/mod.rs
@@ -1,1 +1,0 @@
-pub mod constants;
